% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_cxx.R
\name{check_model}
\alias{check_model}
\title{check_model}
\usage{
check_model(model)
}
\arguments{
\item{model}{list to be checked if it conforms with the model needed in solve_fista_goertler}
}
\value{
nothing if everything is okay
}
\description{
check_model
}
