% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_cxx.R
\name{cxx_estimate_c}
\alias{cxx_estimate_c}
\title{cxx_estimate_c}
\usage{
cxx_estimate_c(model)
}
\arguments{
\item{model}{input model, as constructed by, e.g., empty_model()}
}
\value{
composition (C)
}
\description{
returns estimated phenotype composition, given a trained model. Call cxx implementation.
}
