% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_cxx.R
\name{cxx_evaluate_model}
\alias{cxx_evaluate_model}
\title{cxx_evaluate_model}
\usage{
cxx_evaluate_model(model)
}
\arguments{
\item{model}{input model, as constructed by, e.g., empty_model()}
}
\value{
loss function
}
\description{
returns loss function, evaluated on the given model
}
