% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_FISTA.R
\name{descent_generalized_fista_cxx}
\alias{descent_generalized_fista_cxx}
\title{descent_generalized_fista_cxx}
\usage{
descent_generalized_fista_cxx(model, lambda = 0.01, maxit = 100,
  stop.crit.threshold = 1e-05, save.all.tweaks = FALSE,
  learningrate = NA, linesearchspeed = 2, cycles = 5,
  restarts = TRUE, ...)
}
\arguments{
\item{model}{a model as constructed in interface_cxx.R}

\item{lambda}{a lambda (regularization) parameter}

\item{maxit}{maximum number of iterations}

\item{stop.crit.threshold}{stopping criterion: if loss function changes by no more than this value, assume convergence.}

\item{save.all.tweaks}{if set to TRUE, save all intermediate results (each iteration)}

\item{learningrate}{initial learning rate. if set to NA, determined automatically determined}

\item{linesearchspeed}{rate of learning rate adjustments}

\item{cycles}{number of evaluations in each gradient and nesterov step.}

\item{restarts}{if set to true, restart Nesterov approximation}
}
\value{
a list that contains the trained model and its History
}
\description{
use the cxx implementation of fista.
}
