% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_crossValidation.R
\name{DTD_cv_lambda}
\alias{DTD_cv_lambda}
\title{Cross-validation for digital tissue deconvolution}
\usage{
DTD_cv_lambda(useImplementation = "R", ...)
}
\arguments{
\item{useImplementation}{if set to "R" calls DTD_cv_lambda_R, if set to "cxx" calls DTD_cv_lambda_cxx, passing on all arguments. See the respective implementations of these functions.}
}
\value{
list of length 2.
\itemize{
   \item 'cv.obj', list of lists. DTD model for each lambda, and every folds.
   \item 'best.model', list. DTD model optimized on the complete data set with the best lambda from the cross validation.
}
A cross validation matrix as entry "cv.obj", and the model with minimal loss function
retrained on the complete dataset as "best.model
}
\description{
Our descent generalized FISTA implementation includes a l1 regularization term (see \code{\link{train_deconvolution_model}}.
This function performs a k-fold cross validation to find the best fitting regularization parameter.
For an example see `browseVignettes("DTD")`
This function chooses at runtime either the cxx or the R implementation of cross validation, based on the useImplementation parameter
}
