% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_crossValidation.R
\name{DTD_cv_lambda_cxx}
\alias{DTD_cv_lambda_cxx}
\title{Cross-validation for digital tissue deconvolution}
\usage{
DTD_cv_lambda_cxx(lambda.seq = NULL, tweak.start, X.matrix,
  n.folds = 5, lambda.length = 10, train.data.list,
  cv.verbose = TRUE, warm.start = FALSE, estimate.c.type = "direct",
  ...)
}
\arguments{
\item{lambda.seq}{numeric vector or NULL: Over this series of lambdas the FISTA will be optimized.
If lambda.seq is set to NULL, a generic series of lambdas - depending on the dimensions
of the training set -  will be generated. Default: NULL}

\item{tweak.start}{numeric vector, starting vector for the DTD algorithm.}

\item{X.matrix}{reference matrix, X, each column i holds the gene expression of celltype i.}

\item{n.folds}{integer, number of buckets in the cross validation. Defaults to 10}

\item{lambda.length}{integer, how many lambdas will be generated (only used if lambda.seq is NULL). Defaults to 10}

\item{train.data.list}{list, that can be passed to the F.GRAD.FUN and EVAL.FUN.
Within this list the train/test cross validation will be done.
Notice, that the train.data.list must have an entry named "mixtues". In this entry, the matrix containing the
training samples (as columns) and all features (as rows) must be present. (see Vignette for details)}

\item{cv.verbose}{logical, should information about the cv process be printed to the screen? (Defaults to TRUE)}

\item{warm.start}{logical, should the solution of a previous model of the cross validation be used as a start
in the next model. Defaults to TRUE. Notice, that the warm.start starts with the most unpenalized model.}

\item{estimate.c.type}{string, either "non_negative", or "direct".
Indicates how the algorithm finds the solution of \eqn{arg min_C ||diag(g)(Y - XC)||_2}.
If 'estimate.c.type' is set to "direct" there is no regularization (see \code{\link{estimate_c}}),
if 'estimate.c.type' is set to "non_negative" the estimates "C" must not be negative (non-negative least squares)
(see (see \code{\link{estimate_nn_c}}))}

\item{...}{all parameters that are passed to the \code{\link{descent_generalized_fista}} function.
E.g. maxiter, tweak_vec etc ...}
}
\value{
list of length 2.
\itemize{
   \item 'cv.obj', list of lists. DTD model for each lambda, and every folds.
   \item 'best.model', list. DTD model optimized on the complete data set with the best lambda from the cross validation.
}
A cross validation matrix as entry "cv.obj", and the model with minimal loss function
retrained on the complete dataset as "best.model
}
\description{
Our descent generalized FISTA implementation includes a l1 regularization term (see \code{\link{train_deconvolution_model}}.
This function performs a k-fold cross validation to find the best fitting regularization parameter.
For an example see `browseVignettes("DTD")`
This routine calls the cxx variant of DTD_cv_lambda. See that function for dynamic dispatch, based on the useImplementation parameter
}
