% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_cv_safety_checks.R
\name{DTD_cv_lambda_test_input_generic}
\alias{DTD_cv_lambda_test_input_generic}
\title{input tests for cross validation}
\usage{
DTD_cv_lambda_test_input_generic(
  lambda.seq,
  tweak.start,
  n.folds,
  lambda.length,
  train.data.list,
  cv.verbose,
  warm.start
)
}
\arguments{
\item{lambda.seq}{numeric vector or NULL or "none": Over this series of lambdas the
FISTA will be optimized. If 'lambda.seq' is set to NULL, a generic series of
lambdas - depending on the dimensions of the training set -
will be generated. If 'lambda.seq' is "none", no cross validation is done.
Only one model with lambda = 0 is trained on the complete data set.}

\item{tweak.start}{numeric vector, starting vector for the DTD algorithm.}

\item{n.folds}{integer, number of buckets in the cross validation.}

\item{lambda.length}{integer, how many lambdas will be generated
(only used if lambda.seq is NULL)}

\item{train.data.list}{list, with two entries, a numeric matrix each,
named 'mixtures' and 'quantities'
Within this list the train/test cross validation will be done.
(see Vignette `browseVignettes("DTD")` for details)}

\item{cv.verbose}{logical, should information about the cv process be
printed to the screen?}

\item{warm.start}{logical, should the solution of a previous model of
the cross validation be used as start in the next model.
Notice, that the warm.start starts with the most unpenalized model.

 @return NULL, or it throws an error}
}
\description{
tests common input parameters to the cxx and R implementation of DTD_cv_lambda.
}
