% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_estimateC.R
\name{est.cs}
\alias{est.cs}
\title{Estimating C}
\usage{
est.cs(X, Y, gamma.vec)
}
\arguments{
\item{X}{numeric matrix with cells as columns, and features as rows. Reference matrix X of the DTD problem.}

\item{Y}{numeric matrix with samples as columns, and features as rows.}

\item{gamma.vec}{numeric vector with length of nrow(X). In the equation above the gamma.vec is denotated as g.}
}
\description{
Given a reference matrix X, a matrix of bulks and a g-vector, est.cs finds the solution of
\deqn{arg min || diag(g) (Y - XC) ||_2}
over all C using direct analytical solution
}
\examples{
library(DTD)

# simulate random data:
random.data <- generate.random.data(nTypes = 5,
                                   nSamples.perType = 1,
                                   nFeatures = 100)

# simulate a true c
# (this is not used by the est.cs function, it is only used to show the result!)
true.c <- rnorm(n = ncol(random.data), mean = 1, sd = 0.5)

# calculate bulk y = Xc * some_error
bulk <- random.data \%*\% true.c * rnorm(n = nrow(random.data), mean = 1, sd = 0.01)

#estimate c
estimated.c <- est.cs(X = random.data,
                      Y = bulk,
                      gamma.vec = rep(1, nrow(random.data)))

# visualize that the estimated c are close to the true c
plot(true.c, estimated.c)
}
