% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_estimateCDirect.R
\name{estimate_direct_c}
\alias{estimate_direct_c}
\title{Estimating C}
\usage{
estimate_direct_c(X.matrix, Gamma, new.data)
}
\arguments{
\item{X.matrix}{numeric matrix with cells as columns, and features as rows.
Reference matrix X of the DTD problem. X.matrix can be set to NA (default), if the DTD.model
includes the reference matrix X (default for \code{\link{train_deconvolution_model}})}

\item{Gamma}{either a numeric vector with length of nrow(X),
or a list returned by \code{\link{train_deconvolution_model}}, \code{\link{DTD_cv_lambda}},
or\code{\link{descent_generalized_fista}}. In the equation above
  the DTD.model provides the vector g.}

\item{new.data}{numeric matrix with samples as columns, and features as rows.
In the formula above denoated as Y.}
}
\value{
numeric matrix with ncol(X) rows, and ncol(Y) columns
}
\description{
Given a reference matrix X, a matrix of bulks Y and a g-vector, "estimate_c" finds
the solution of \deqn{arg min || diag(g) (Y - XC) ||_2} over all C using
direct analytical solution: (see Goertler et al. 2018)
\deqn{ C(g) = (X^T \Gamma X )^(-1) X^T \Gamma Y}
with \eqn{\Gamma} = diag(g)
}
