% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_estimateCNN.R
\name{estimate_nn_c}
\alias{estimate_nn_c}
\title{Estimating non negative C}
\usage{
estimate_nn_c(X.matrix = NA, new.data, Gamma)
}
\arguments{
\item{X.matrix}{numeric matrix, with features/genes as rows,
and cell types as column. Each column of X.matrix is a reference
expression profile}

\item{new.data}{numeric matrix with samples as columns,
and features/genes as rows. In the formula above denoated as Y.}

\item{Gamma}{numeric matrix, with nrow(X) x nrow(X)}
}
\value{
numeric matrix with ncol(X) rows, and ncol(Y) columns
}
\description{
Given a reference matrix X, a matrix of bulks Y and a g-vector,
"estimate_nn_c" finds the solution of
\deqn{arg min || diag(g) (Y - XC) ||_2} with non negative constraint over C,
using the package 'nnls' (non-negative least squares)
}
