% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_eval_cor.R
\name{evaluate_cor}
\alias{evaluate_cor}
\title{Evaluate correlation}
\usage{
evaluate_cor(X, Y, C, tweak)
}
\arguments{
\item{X}{}

\item{Y}{}

\item{C}{}

\item{tweak}{}
}
\description{
The loss-function learning digital tissue deconvolution finds a vector g which minimizes the Loss function L\cr
\deqn{L = \sum cor(true_C, estimatd_C(g))}
For a given vector g the evaluate_cor function returns the value of the loss function.
}
