% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plotConvergence.R
\name{ggplot_correlation}
\alias{ggplot_correlation}
\title{Plot loss curve}
\usage{
ggplot_correlation(fista.output, test.set)
}
\arguments{
\item{test.set}{}
}
\description{
The "ggplot_correlation" function uses ggplot2 and reshape to visualize the decrease of the loss-function.
}
\details{
As input parameter it needs the output of \code{\link{descent_generalized_fista}}. If fista has been evoked
with save_all_tweaks = T, and a test set is available, the loss-function can be evaluated for all tweak vectors,
and both training and test loss curves will be plotted (=> detect overfitting).\cr
For data generation, and evoking the FISTA algorithm see \code{\link{descent_generalized_fista}}.
}
\examples{
# the following example requires the data generated in the example of descent_generalized_fista.
ggplot_correlation(fista.output = catch, test.set = NA)
}
