% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plotCV.R
\name{ggplot_cv}
\alias{ggplot_cv}
\title{Visualizing cross validation output}
\usage{
ggplot_cv(crossValFrame, main = "", LAMBDA.TRANS.FUN = log2)
}
\arguments{
\item{LAMBDA.TRANS.FUN}{}
}
\description{
Visualizing cross validation output
}
