% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plot_explCor.R
\name{ggplot_explained_correlation}
\alias{ggplot_explained_correlation}
\title{Plot the explained correlation per feature}
\usage{
ggplot_explained_correlation(DTD.model, X.matrix = NA, test.data,
  estimate.c.type, print.labels = FALSE)
}
\arguments{
\item{DTD.model}{list as returned by \code{\link{train_correlatio_model}}, \code{\link{DTD_cv_lambda}},
or \code{\link{descent_generalized_fista}}}

\item{X.matrix}{numeric matrix, with features/genes as rows, and cell types as column.
Each column of X.matrix is a reference expression profile}

\item{test.data}{numeric matrix with samples as columns, and features as rows.
In the deconvolution formula 'test.data' is denoated as Y.}

\item{estimate.c.type}{string, either "non_negative", or "direct". Indicates how the algorithm finds the solution of
\eqn{arg min_C ||diag(g)(Y - XC)||_2}. If estimate.c.type is set to "direct" there is no regularization
(see \code{\link{estimate_c}}),
if estimate.c.type is set to "non_negative" the estimates "C"
must not be negative (non-negative least squares) (see (see \code{\link{estimate_nn_c}}))}

\item{print.labels}{logical, should the feature names be plotted into the picture? (Defaults to FALSE)}
}
\value{
ggplot object
}
\description{
In order to assess the importance of a feature in the deconvolution process, we can exclude it from a trained model,
and observe the change of correlaion on a test set. If the correlation e.g. decreases by 1 %,
the gene explains 1 % correlation within the deconvolution model.
The 'ggplot_explained_correlation' function iteratively excludes each feature from the trained model,
and visualizes the explained correlation.
}
\details{
For an example see section "Explained correlation" in the package vignette `browseVignettes("DTD")`
}
