% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plot_heatmap.R
\name{ggplot_heatmap}
\alias{ggplot_heatmap}
\title{Heatmap of diag(g) * X}
\usage{
ggplot_heatmap(DTD.model, X.matrix = NA, test.data = NULL,
  estimate.c.type, main = "", feature.subset = 100)
}
\arguments{
\item{DTD.model}{list as returned by \code{\link{train_correlatio_model}}, \code{\link{DTD_cv_lambda}},
or \code{\link{descent_generalized_fista}}}

\item{X.matrix}{numeric matrix, with features/genes as rows, and cell types as column.
Each column of X.matrix is a reference expression profile}

\item{test.data}{numeric matrix with samples as columns, and features as rows.
In the deconvolution formula 'test.data' is denoated as Y.}

\item{estimate.c.type}{string, either "non_negative", or "direct". Indicates how the algorithm finds the solution of
\eqn{arg min_C ||diag(g)(Y - XC)||_2}. If estimate.c.type is set to "direct" there is no regularization
(see \code{\link{estimate_c}}),
if estimate.c.type is set to "non_negative" the estimates "C"
must not be negative (non-negative least squares) (see (see \code{\link{estimate_nn_c}}))}

\item{main}{string, used as title in the plot}

\item{feature.subset}{numeric or a vector of strings. If it is a numeric, "subset" features will be picked from the
'explained correlation' ranking (if 'feature.subset' < 1, this is the fraction of feature, if 'feature.subset' > 1
it is the total amount. If it is a vector of strings, these genes will be used.}
}
\value{
ggplot object
}
\description{
Plot a heatmap of the reference matrix X times , using a selected subset of features
NOT FINISHED, STOP
Top features are selected via "explained correlation":
In order to assess the importance of a feature in the deconvolution process, we can exclude it from a trained model,
and observe the change of correlaion on a test set. If the correlation e.g. decreases by 1 %,
the gene explains 1 % correlation within the deconvolution model.
The 'ggplot_heatmap' function iteratively excludes each feature from the trained model,
resulting in a ranking for the genes.
The clustering for the heatmp will be calculated using only these features.
}
\details{
For an example see section "Explained correlation" in the package vignette `browseVignettes("DTD")`
}
