% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_gradient.R
\name{gradient_cor_trace}
\alias{gradient_cor_trace}
\title{Gradient of \eqn{L(g) = \sum cor(true_C, estimated_C(g) )}}
\usage{
gradient_cor_trace(X, Y, C, tweak, estimate.c.type)
}
\arguments{
\item{Y}{numeric matrix with samples as columns,
and features/genes as rows. Holding bulk gene expressions}

\item{C}{numeric matrix with reference samples as rows,
and samples as columns. Holding cellular composionts}

\item{tweak}{numeric vector with length of nrow(X).
In the Loss function above tweak is named "g"}

\item{estimate.c.type}{string, either "non_negative", or "direct".
Indicates how the algorithm finds the solution of
\eqn{arg min_C ||diag(g)(Y - XC)||_2}.
\itemize{
   \item If 'estimate.c.type' is set to "direct",
 there is no regularization (see \code{\link{estimate_c}}),
   \item if 'estimate.c.type' is set to "non_negative",
 the estimates "C" must not be negative (non-negative least squares)
(see (see \code{\link{estimate_nn_c}}))
}}

\item{X.matrix}{numeric matrix, with features/genes as rows,
and cell types as column. Each column of X.matrix is a reference
expression profile}
}
\value{
numeric list, same length as "tweak"
}
\description{
This function returns the value of the gradient of our Loss-function L.\cr
For the mathematical theory see Goertler et al, 2018.
For examples see code of \code{\link{train_deconvolution_model}}
}
