% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_cv_lambda_generator.R
\name{lambda_sequence}
\alias{lambda_sequence}
\title{generates a sequence of lambdas}
\usage{
lambda_sequence(lambda.seq, lambda.length, train.Y)
}
\arguments{
\item{lambda.seq}{sequence of lambdas. may be null.}

\item{lambda.length}{number of lambdas to generate}

\item{train.Y}{training data, used for heuristics involving the dimensionality of the problem}
}
\value{
sequence of lambdas (numeric vector)
}
\description{
depending on the input, generates a sequence of lambdas or makes it usable (sorts it)
}
