% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_cv_bucket_maker.R
\name{make_buckets}
\alias{make_buckets}
\title{map every sample to a bucket}
\usage{
make_buckets(train.Y, folds)
}
\arguments{
\item{train.Y}{training data}

\item{folds}{number of buckets to build.}
}
\value{
list of lists of same structure as train.Y, but with names indicating the bucket.
}
\description{
takes training data and sorts it into buckets, one for each fold.
}
