% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_mix_with_Jitter.R
\name{mix.samples.jitter}
\alias{mix.samples.jitter}
\title{Mix samples with Jitter}
\usage{
mix.samples.jitter(sample.names, special.samples, nSamples = 1000, datamatrix,
  pheno, verbose = FALSE, singleSpecial = FALSE, add_jitter = FALSE,
  chosen.mean = 1, chosen.sd = 0.05, min.amount.samples = 1,
  per.type = 1, included.in.X)
}
\arguments{
\item{sample.names}{vector of strings, have to match colnames(datamatrix).}

\item{special.samples}{vector of strings, have to match colnames(datamatrix), these samples will
occur with higher quantites within the mix data.}

\item{nSamples}{integer, how many mixtures should be drawn}

\item{datamatrix}{numeric matrix, with samples as columns, and features as rows}

\item{pheno}{named list of characters, indicates which of the samples in datamatrix belongs to which type in sample.names/special.names}

\item{verbose}{boolean, should function tell about progression?}

\item{singleSpecial}{boolean, should all special names be used? or only a single one}

\item{add_jitter}{boolean, should the mixtures be mulitplied with a vector of normally distributed numbers? (JITTER)}

\item{chosen.mean}{numeric, mean of jitter}

\item{chosen.sd}{numeric, standard deviation of jitter}

\item{min.amount.samples}{integer, how many samples have to be present such that it averages over them,
instead of taking only 1}

\item{per.type}{integer, how many samples "per type" should be used for each mixture}

\item{included.in.X}{list of strings, which cell types are included in the reference matrix.
The mix function will mix all samples in the data set, but will only return the quantity matrix
for the types included in X (and in the right ordering)}
}
\value{
list with two entries. "quantities" matrix (nrow = ncol(datamatrix), ncol = nMixtures) and "mixture"
matrix (nrow = nrow(datamatrix), ncol = nMixtures)
}
\description{
"mix.samples.jitter" takes pheno information (sample.names, special.names) and a expression matrix.
Its output is a expression and a quantity matrix.
Each sample of the output is a mixture of input samples, multiplied with Jitter vector.
"mix.samples.Jitter" mixes samples in a way that they look similar to biological data.
For example, in a tumor tissue there are several cells included, but we expect that
most of them are tumor cells. Therefore, the mix.samples.Jitter function expects a
list of samples which represent immune cells (and occur in minor fractions) and a
list of special samples (which occur in major fractions)
}
\examples{
library(DTD)
random.data <- generate.random.data(nTypes = 10,
                                    nSamples.perType = 10,
                                    nFeatures = 500,
                                    sample.type = "Cell",
                                    feature.type = "gene")

# normalize all samples to the same amount of counts:
random.data <- normalizeToCount(random.data)

# extract indicator list.
# This list contains the type of the sample as value, and the sample name as names
indicator.list <- gsub("^Cell[0-9]*\\\\.", "", colnames(random.data))
names(indicator.list) <- colnames(random.data)

# extract reference matrix X
# First, decide which cells should be deconvoluted.
# Notice, in the mixtures there can be more cells than in the reference matrix.
include.in.X <- c("Type2", "Type3", "Type4", "Type5")

X.matrix <- matrix(NA, nrow=nrow(random.data), ncol=length(include.in.X))
colnames(X.matrix) <- include.in.X
rownames(X.matrix) <- rownames(random.data)

percentage.of.all.cells <- 0.2

# samples that are included in X must not be used in the training set!
samples.to.remove <- c()

for(l.type in include.in.X){
  all.of.type <- names(indicator.list)[which(indicator.list == l.type)]
  chosen.for.X <- sample(x = all.of.type,
                         size = length(all.of.type) * percentage.of.all.cells,
                         replace = FALSE)
  samples.to.remove <- c(samples.to.remove, chosen.for.X)

  average <- rowSums(random.data[, samples.to.remove])
  X.matrix[, l.type] <- average
}

# here, I declare "Type1" as Tumor cells, and all other as immune cells
special.samples <- c("Type1")
all.samples <- unique(indicator.list)
sample.names <- all.samples[- which(all.samples \%in\% special.samples)]

training.data <- random.data[ , -which(colnames(random.data) \%in\% samples.to.remove)]

training.data <- mix.samples.jitter(sample.names = sample.names,
                                     special.samples = special.samples,
                                     nSamples = 1e3,
                                     datamatrix = random.data,
                                     pheno = indicator.list,
                                     singleSpecial = FALSE,
                                     add_jitter = TRUE,
                                     chosen.mean = 1,
                                     chosen.sd = 0.05,
                                     min.amount.samples = 1)
}
