% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_norms.R
\name{n2normed}
\alias{n2normed}
\title{Norm a vector to ||x||_2 = 1}
\usage{
n2normed(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector with same length as x
}
\description{
Norm a vector to ||x||_2 = 1
}
\examples{
set.seed(1)
same.values <- rnorm(n=10)
norm(same.values, type = "2")
normed.values <- n2normed(same.values)
norm(normed.values, type = "2")

}
