% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_nesterovFactors.R
\name{nesterov_factor}
\alias{nesterov_factor}
\title{Nesterov Factor 2/(k+1)}
\usage{
nesterov_factor(k)
}
\arguments{
\item{k}{integer, iteration step of the FISTA algorithm}
}
\value{
integer
}
\description{
The FISTA algorithm extrapolates/correctes after the gradient step. The "nesterov_factor" function
returns how far the FISTA will exptrapolate in a given step "k"
}
\examples{
nesterov_factor(2)
}
