% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_nesterovFaktors.R
\name{nesterov_faktor}
\alias{nesterov_faktor}
\title{Nesterov Faktor 2/(k+1)}
\usage{
nesterov_faktor(k)
}
\arguments{
\item{k}{integer, iteration step of the FISTA algorithm}
}
\value{

}
\description{
The FISTA algorithm extrapolates/correctes after the gradient step. The "nesterov_faktor" function
 returns how far the FISTA will exptrapolate in a given step "k"
}
\examples{
nesterov_faktor(2)
}
