% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_cv_pick_mean_test.R
\name{pick.mean.test.results.function}
\alias{pick.mean.test.results.function}
\title{returns mean of the individual test results, setting unconverged or unavailable results to Inf.}
\usage{
pick.mean.test.results.function(lambda.list)
}
\arguments{
\item{lambda.list}{output of a DTD cross validation run (list of lists)}
}
\value{
list of means.
}
\description{
returns mean of the individual test results, setting unconverged or unavailable results to Inf.
}
