% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_nesterovFaktors.R
\name{positive_subspace_pmax}
\alias{positive_subspace_pmax}
\title{Wrapper for pmax}
\usage{
positive_subspace_pmax(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
pmax(x,0), numeric vector with same length as x, but without any negative entries
}
\description{
positive_subspace_pmax expects a vector x, and applies pmax(x, 0) on it.
It is needed as a wrapper within the \code{\link{descent_generalized_fista}} implementation.
}
\examples{
set.seed(2202)
vec.with.neg <- rnorm(10)
range(vec.with.neg)
vec.without.neg <- positive_subspace_pmax(vec.with.neg)
range(vec.without.neg)
}
