% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_sampleX.R
\name{sample_random_X}
\alias{sample_random_X}
\title{Sample random X}
\usage{
sample_random_X(included.in.X, pheno, exp.data,
  percentage.of.all.cells = 0.1)
}
\arguments{
\item{included.in.X}{vector of strings, which cell types should be included in X?}

\item{pheno}{named vector of strings, names have to match exp.data.
Information about the cell type (values of vector) for each sample (names of vector)}

\item{exp.data}{numeric matrix with features as rows, and samples as columns}

\item{percentage.of.all.cells}{float, which percentage of all possible cells should
be use to generate a cell type profile? Defaults to 0.1}
}
\value{
list with two entries: (1) X.matrix: numeric matrix with as many rows as exp.data,
and as many columns as length(included.in.X)
(2) samples.to.remove: vector of strings, all samples that have been used in generating X.
}
\description{
The loss-function learning digital tissue deconvolution approach published by Goertler et al 2018
estimates cell compositions for a given reference matrix X. Basically, there are two methods to
specify the reference profiles in X. Either they are selected using external knowledge
(e.g. additional measurements) or they are randomly selected out of the complete data set.
The sample_random_X function is an implementation for the second method.
}
\details{
For examples see the DTD vignette: browseVignettes("DTD")
}
