% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_cv_select_subset.R
\name{select.fun}
\alias{select.fun}
\title{selects a specific sample from a list of entries}
\usage{
select.fun(list.entry, samples)
}
\arguments{
\item{list.entry}{list of entries (matrix or list)}

\item{samples}{samples, serving as an index to list.entry.}
}
\value{
specific entry of list.entry
}
\description{
selects a specific sample from a list of entries
}
