% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_cxx.R
\name{set_model_coeff_estimation}
\alias{set_model_coeff_estimation}
\title{set_model_coeff_estimation}
\usage{
set_model_coeff_estimation(model, coeffestimname)
}
\arguments{
\item{model}{input model, as constructed by, e.g., empty_model()}

\item{coeffestimname}{the desired value for the coefficient estimation function. May currently be either "direct" or "nnls" (for non-negative least squares).}
}
\value{
the updated model with the coefficient estimation function set.
}
\description{
set_model_coeff_estimation
}
