% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_cxx.R
\name{set_model_inversion_precision}
\alias{set_model_inversion_precision}
\title{set_model_inversion_precision}
\usage{
set_model_inversion_precision(model, precision)
}
\arguments{
\item{model}{input model, as constructed by, e.g., empty_model()}

\item{precision}{positive number against which the norm is compared to and the program crashes if this value is exceeded.}
}
\value{
the updated model with the precision set.
}
\description{
set_model_inversion_precision
}
