% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_cxx.R
\name{set_model_normfunction}
\alias{set_model_normfunction}
\title{set_model_normfunction}
\usage{
set_model_normfunction(model, normfnname)
}
\arguments{
\item{model}{input model, as constructed by, e.g., empty_model()}

\item{normfnname}{the desired value for the norm function. May currently be either "identity" or "norm2"}
}
\value{
the updated model with the norm function set.
}
\description{
set_model_normfunction
}
