% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_cxx.R
\name{set_model_subspacefunction}
\alias{set_model_subspacefunction}
\title{set_model_subspacefunction}
\usage{
set_model_subspacefunction(model, subspfnname)
}
\arguments{
\item{model}{input model, as constructed by, e.g., empty_model()}

\item{subspfnname}{the desired value for the subspace function. May currently be only "positive".}
}
\value{
the updated model with the subspace function set.
}
\description{
set_model_subspacefunction
}
