% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_softThresholding.R
\name{soft_thresholding}
\alias{soft_thresholding}
\title{Soft thresholding}
\usage{
soft_thresholding(x, lambda)
}
\arguments{
\item{x}{list of numerics, x = g - step.size * gradient(g)}

\item{lambda}{float or list of floats, regularization parameter}
}
\value{
vector of numerics, same length as x
}
\description{
implementation of the proximal operator for l1 penalty
(see e.g. Bubeck 2015)
}
\examples{
set.seed(1)
soft_thresholding(x = runif(10) - rnorm(10), lambda = 0.2)
}
