% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_cxx.R
\name{solve_fista_goertler}
\alias{solve_fista_goertler}
\title{Title}
\usage{
solve_fista_goertler(model, lambda = 0.01, maxiter = 100,
  stop.crit.threshold = 1e-05, save.all.tweaks = FALSE,
  learningrate = NA, linesearchspeed = 2, cycles = 5,
  restarts = TRUE)
}
\arguments{
\item{restarts}{}
}
\description{
Title
}
