% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{test_c_type}
\alias{test_c_type}
\title{test_c_type
(All test functions are used for safety checks in the DTD package)}
\usage{
test_c_type(test.value, output.info)
}
\arguments{
\item{test.value}{value to be tested for integer}

\item{output.info}{[1] function that calls, [2] name of value to test}
}
\value{
NULL or stops with an error
}
\description{
test_c_type
(All test functions are used for safety checks in the DTD package)
}
