% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{test_integer}
\alias{test_integer}
\title{test_integer
(All test functions are used for safety checks in the DTD package)}
\usage{
test_integer(test.value, output.info, min, max)
}
\arguments{
\item{test.value}{value to be tested for integer}

\item{output.info}{[1] function that calls, [2] name of value to test}

\item{min}{min is tested '<'}

\item{max}{max is tested '>'}
}
\value{
TRUE, if no error is detected, stops with error otherwise
}
\description{
test_integer
(All test functions are used for safety checks in the DTD package)
}
