% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{test_logical}
\alias{test_logical}
\title{test_logical
(All test functions are used for safety checks in the DTD package)}
\usage{
test_logical(test.value, output.info)
}
\arguments{
\item{test.value}{value to be tested for integer}

\item{output.info}{[1] function that calls, [2] name of value to test}
}
\value{
TRUE, or it throws an error
}
\description{
test_logical
(All test functions are used for safety checks in the DTD package)
}
