% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{test_numeric}
\alias{test_numeric}
\title{test_numeric
(All test functions are used for safety checks in the DTD package)}
\usage{
test_numeric(test.value, output.info, min, max)
}
\arguments{
\item{test.value}{value to be tested for integer}

\item{output.info}{[1] function that calls, [2] name of value to test}

\item{min.value}{min is tested '<'}

\item{max.value}{max is tested '>'}
}
\value{
TRUE, if no error has occured, stops with error otherwise
}
\description{
test_numeric
(All test functions are used for safety checks in the DTD package)
}
