% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{test_string}
\alias{test_string}
\title{test_string
(All test functions are used for safety checks in the DTD package)}
\usage{
test_string(test.value, output.info)
}
\arguments{
\item{test.value}{value to be tested for 'as.character' usage}

\item{output.info}{[1] function that calls, [2] name of value to test}
}
\value{
string
}
\description{
test_string
(All test functions are used for safety checks in the DTD package)
}
