% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{test_tweak_vec}
\alias{test_tweak_vec}
\title{test_tweak_vec
(All test functions are used for safety checks in the DTD package)}
\usage{
test_tweak_vec(tweak.vec, output.info)
}
\arguments{
\item{tweak.vec}{vector which will be tested for numeric and NAs}

\item{output.info}{[1] function that calls, [2] name of value to test}
}
\value{
TRUE, if no error has occured, stops with error otherwise
}
\description{
test_tweak_vec
(All test functions are used for safety checks in the DTD package)
}
