% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_gradient.R
\name{Trace.H.gradient}
\alias{Trace.H.gradient}
\title{Gradient of \eqn{L(g) = \sum cor(true_C, estimated_C(g) }}
\usage{
Trace.H.gradient(X = x_mat, Y = y_mat, C = c_mat, gamma.vec = v_vec)
}
\arguments{
\item{X}{numeric matrix with features as rows, and reference samples as columns}

\item{Y}{numeric matrix with features as rows, and samples as columns}

\item{C}{numeric matrix with reference samples as rows, and samples as columns}

\item{gamma.vec}{numeric vector with length of nrow(X).
In the Loss function above gamma.vec is named "g"}
}
\value{
numeric list, same length as "gamma.vec"
}
\description{
This function returns the value of the gradient of our Loss-function L.
}
\examples{

}
