% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_estimateCDirect.R
\name{estimate_direct_c}
\alias{estimate_direct_c}
\title{Estimating C}
\usage{
estimate_direct_c(X.matrix, Gamma, new.data)
}
\arguments{
\item{X.matrix}{numeric matrix, with features/genes as rows,
and cell types as column. Each column of X.matrix is a reference
expression profile.}

\item{Gamma}{numeric matrix, with nrow(X) x nrow(X)}

\item{new.data}{numeric matrix with samples as columns,
and features/genes as rows. In the formula above denoated as Y.}
}
\value{
numeric matrix with ncol(X) rows, and ncol(Y) columns
}
\description{
Given a reference matrix X, a matrix of bulks Y and a g-vector, "estimate_c" finds
the solution of \deqn{arg min || diag(g) (Y - XC) ||_2} over all C using
direct analytical solution: (see Goertler et al. 2018)
\deqn{ C(g) = (X^T \Gamma X )^(-1) X^T \Gamma Y}
with \eqn{\Gamma} = diag(g)
}
