% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_generate_random_data.R
\name{generate.random.data}
\alias{generate.random.data}
\title{Generating random data}
\usage{
generate.random.data(nTypes = 5, nSamples.perType = 10, nFeatures = 1000,
  sample.type = "Cell", feature.type = "gene", seed = 1310)
}
\arguments{
\item{nTypes}{integer, how many different types should be included in the data set}

\item{nSamples.perType}{integer, how many samples should be generated per type.
Notice, for each type, the number of samples will be randomized a bit.}

\item{nFeatures}{integer, how many features should be included}

\item{sample.type}{string, name of samples}

\item{feature.type}{string, name of features}

\item{seed}{integer, to which the seed will be set}
}
\value{
matrix with (nTypes * nSamples.perType) columns, and nFeatures rows
}
\description{
generate.random.data simulates data which can be used exemplary for digital tissue deconvolution.
It will generate a numeric matrix with "nFeatures" rows, and "nTypes" * "nSamples.perType"
columns. Each column represents a sample of special type. The function will generate "nTypes",
and for each type "nSamples.perType".
}
\examples{
 library(DTD)
random.data <- generate.random.data(nTypes = 5,
                                    nSamples.perType = 10,
                                    nFeatures = 100,
                                    sample.type = "Cell",
                                    feature.type = "gene")

# undo log transformation to have counts:

random.data <- (2^random.data) - 1

# normalize all samples to the same amount of counts:
random.data <- normalizeToCount(random.data)
}
