% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plotConvergence.R
\name{ggplot_convergence}
\alias{ggplot_convergence}
\title{Plot loss curve}
\usage{
ggplot_convergence(DTD.model, EVAL.FUN = NA, main = "")
}
\arguments{
\item{DTD.model}{list, as returned by the train_correlation_model function.}

\item{EVAL.FUN}{function, that takes a single input (the tweak vector of the fista.output),
and returns the loss function on the test set. Default is NA}

\item{main}{string, additionally title (default "")}
}
\value{
ggplot object
}
\description{
The "ggplot_convergence" function uses ggplot2 and reshape to visualize the decrease of the loss-function after
a model has been trained.
}
\details{
As input parameter it needs the output of \code{\link{train_correlatio_model}}, \code{\link{DTD_cv_lambda}},
or\code{\link{descent_generalized_fista}}.
If fista has been evoked with save_all_tweaks = T, and a evaluation function is available,
the loss-function can be evaluated for all intermediate tweak vectors,
and both training and test loss curves will be plotted (=> detect overfitting).\cr
For an example see section "Visualization of learn curve" in the package vignette `browseVignettes("DTD")`
}
