% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plotConvergence.R
\name{ggplot_correlation}
\alias{ggplot_correlation}
\title{Plot loss curve}
\usage{
ggplot_correlation(fista.output, test.set = NA, X.matrix = NA, main = "")
}
\arguments{
\item{fista.output}{list, with "Convergence" entry as return by the descent_generalized_fista function}

\item{test.set}{list with "mixtures" matrix, and "quantity" matrix as returned by mix.samples or mix.samples.jitter function}

\item{X.matrix}{numeric matrix, reference matrix of the DTD problem}

\item{main}{string, additionally title}
}
\value{
ggplot object
}
\description{
The "ggplot_correlation" function uses ggplot2 and reshape to visualize the decrease of the loss-function after
a model has been trained.
}
\details{
As input parameter it needs the output of \code{\link{descent_generalized_fista}}. If fista has been evoked
with save_all_tweaks = T, and a test set is available, the loss-function can be evaluated for all tweak vectors,
and both training and test loss curves will be plotted (=> detect overfitting).\cr
For an example see
\code{\link{descent_generalized_fista}}
}
