% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plotCV.R
\name{ggplot_cv}
\alias{ggplot_cv}
\title{Visualizing cross validation output}
\usage{
ggplot_cv(DTD.model, title = "", LAMBDA.TRANS.FUN = log2,
  upper.x.axis.info = "non-zero", x.lab = expression("log2(" * lambda *
  ")"), y.lab = "Loss function")
}
\arguments{
\item{DTD.model}{: list as returned by \code{\link{train_deconvolution_model}}
\code{\link{DTD_cv_lambda_R}} or \code{\link{DTD_cv_lambda_cxx}}}

\item{title}{string, additionally title}

\item{LAMBDA.TRANS.FUN}{function, will be applied to the lambda sequence,
e.g. to get equidistance x ticks}

\item{upper.x.axis.info}{string, either "non-zero" or "geometric-mean".
Sets the information that will be printed on the upper x axis.}

\item{x.lab}{string, used as x label on the plot}

\item{y.lab}{string, used as y label on the plot}
}
\value{
a ggplot plot object
}
\description{
Plot the regularization parameter lambda vs the resulting loss function value.
The lambda with the minimal mean loss in cross validation is marked with a
red dot. For an example see "Cross validation" in the package vignette
`browseVignettes("DTD")`
}
