% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plotTweakHist.R
\name{ggplot_ghistogram}
\alias{ggplot_ghistogram}
\title{Histogram of the g/tweak vector}
\usage{
ggplot_ghistogram(fista.output, n.bins = 50,
  TRANSFORM.FUN = DTD::identity, main = "", x.lab = "g-vec")
}
\arguments{
\item{fista.output}{: list with "Tweak" entry. The result of a \code{\link{descent_generalized_fista}} call}

\item{n.bins}{: integer, number of bins in the histogram. Defaults to 50}

\item{TRANSFORM.FUN}{: function that expects a list of floats, and returns a list of floats
Defaults to identity.}
}
\value{
ggplot object
}
\description{
The resulting distribution of the entries of g are visualized best in a histogram.
As the g vector may get spread over a big range, you can provide a transformation function.
This will be applied on the g/tweak vector. If the provided transformation function returns NA,
the function will state the number of NAs in the title of the plot.
For an example see `browseVignettes("DTD")`
}
