% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plotTweakHist.R
\name{ggplot_ghistogram}
\alias{ggplot_ghistogram}
\title{Histogram of the g/tweak vector}
\usage{
ggplot_ghistogram(DTD.model, n.bins = 50,
  TRANSFORM.FUN = DTD::identity, title = "", x.lab = "g-vec")
}
\arguments{
\item{DTD.model}{either a numeric vector ,
or a list returned by \code{\link{train_deconvolution_model}}, \code{\link{DTD_cv_lambda}},
or\code{\link{descent_generalized_fista}}.}

\item{n.bins}{: positive integer, number of bins in the histogram. Defaults to 50}

\item{TRANSFORM.FUN}{: function that expects a list of floats, and returns a list of floats
Defaults to identity. (defaults to identity)}

\item{title}{string, additionally title (default "")}

\item{x.lab}{string, used as x label of the plot. Defaults to ""}
}
\value{
ggplot object
}
\description{
The resulting distribution of g are visualized best in a histogram.
As the g vector may get spread over a big range, you can provide a transformation function.
This will be applied on the g/tweak vector. If the provided transformation function returns NA,
the function will state the number of NAs in the title of the plot.
For an example see section "Histogram of g-vector" in the package vignette `browseVignettes("DTD")`
}
