% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plot_heatmap.R
\name{ggplot_heatmap}
\alias{ggplot_heatmap}
\title{clustered heatmap of diag(g) * X}
\usage{
ggplot_heatmap(DTD.model, X.matrix = NA, test.data = NULL,
  estimate.c.type, title = "", feature.subset = 100,
  log2.expression = TRUE)
}
\arguments{
\item{DTD.model}{list as returned by \code{\link{train_deconvolution_model}}, \code{\link{DTD_cv_lambda}},
or \code{\link{descent_generalized_fista}}, or a numeric vector, which will be used as 'g'.}

\item{X.matrix}{numeric matrix, with features/genes as rows, and cell types as column.
Each column of X.matrix is a reference expression profile.
Notice, if the model includes a 'reference.X' matrix, this parameter can be set to NA.}

\item{test.data}{numeric matrix with samples as columns, and features as rows.
In the deconvolution formula 'test.data' is denoated as Y.
If subsetting is done via feature names, 'test.data' will not be used.}

\item{estimate.c.type}{string, either "non_negative", or "direct". Indicates how the algorithm finds the solution of
\eqn{arg min_C ||diag(g)(Y - XC)||_2}. \cr
If estimate.c.type is set to "direct" there is no regularization (see \code{\link{estimate_c}}),\cr
if estimate.c.type is set to "non_negative" the estimates "C"
must not be negative (non-negative least squares) (see (see \code{\link{estimate_nn_c}}))}

\item{title}{string, additionally title (default "")}

\item{feature.subset}{numeric or a vector of strings. If it is a numeric, "subset" features will be picked from the
'explained correlation' ranking (if 'feature.subset' < 1, this is the fraction of feature, if 'feature.subset' > 1
it is the total amount). If it is a vector of strings, these features will be used.}

\item{log2.expression}{logical, in the heatmap, should the values be log transformed?}
}
\value{
ggplot object
}
\description{
For a DTD.model the 'ggplot_heatmap' function visualizes
\deqn{diag(g) * X} on a subset of features as a clusterd heatmap.\cr
Feature subsetting can either be done by a vector of strings, that match the feature names of X.\cr
or via 'explained correlation':
In order to assess the importance of a feature in the deconvolution process, we can exclude it from a trained model,
and observe the change of correlaion on a test set. If the correlation e.g. decreases by 1 %,
the gene explains 1 % correlation within the deconvolution model.
The 'ggplot_heatmap' function iteratively excludes each feature from the trained model,
resulting in a ranking for the genes.
The clustering for the heatmp will be calculated using only these features.
}
\details{
For an example see section "Explained correlation" in the package vignette `browseVignettes("DTD")`
}
