% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plot_esti_vs_true.R
\name{ggplot_true_vs_esti}
\alias{ggplot_true_vs_esti}
\title{Plot true vs estimated cell composition}
\usage{
ggplot_true_vs_esti(estimatedC, trueC, norm.columnwise = TRUE,
  title = "", color.indi = NA)
}
\arguments{
\item{estimatedC}{matrix, with cell types as rows, and mixtures as columns}

\item{trueC}{matrix, with cell types as rows, and mixtures as columns}

\item{norm.columnwise}{boolean, should every column be normalized to sum of 1?}

\item{title}{string, title for plot}

\item{color.indi}{vector with length of ncol(trueC), used for colouring in ggplot}
}
\value{
ggplot object
}
\description{
'ggplot_true_vs_esti' can be used to evaluate a trained DTD model,
by plotting known/true quantities versus estimated quantities per cell type. Its main inputs are two quantity matrices.
For an example see
\code{\link{descent_generalized_fista}}
}
