% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_gradient.R
\name{gradient_cor_trace}
\alias{gradient_cor_trace}
\title{Gradient of \eqn{L(g) = \sum cor(true_C, estimated_C(g) )}}
\usage{
gradient_cor_trace(X, Y, C, tweak)
}
\arguments{
\item{X}{numeric matrix with features as rows, and reference samples as columns}

\item{Y}{numeric matrix with features as rows, and samples as columns}

\item{C}{numeric matrix with reference samples as rows, and samples as columns}

\item{tweak}{numeric vector with length of nrow(X).
In the Loss function above tweak is named "g"}
}
\value{
numeric list, same length as "tweak"
}
\description{
This function returns the value of the gradient of our Loss-function L.\cr
For the mathematical theory see Goertler et al, 2018.
For examples see code of \code{\link{train_correlation_model}}
}
