% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_cv_bucket_maker.R
\name{make_buckets}
\alias{make_buckets}
\title{map every sample to a bucket}
\usage{
make_buckets(train.Y, folds)
}
\arguments{
\item{train.Y}{numeric matrix, training data}

\item{folds}{integer, number of buckets to build.}
}
\value{
named vector
}
\description{
takes training data and sorts it into buckets, one for each fold.
}
\examples{
some.mat <- matrix(1:20, nrow = 2)
colnames(some.mat) <- paste0("some.mat", 1:ncol(some.mat))
n.folds <- 5
mapped.buckets <- make_buckets(
  train.Y = some.mat
  , folds = n.folds
  )
table(mapped.buckets)
}
