% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_normToCount.R
\name{normalize_to_count}
\alias{normalize_to_count}
\title{normalize_to_count}
\usage{
normalize_to_count(exp.data, count = 1e+06)
}
\arguments{
\item{exp.data}{positive numeric matrix, with samples as columns and features as columns}

\item{count}{float , to which every sample of exp.data will be scaled}
}
\value{
"ret", matrix with same dimension as "exp.data"
}
\description{
'normalize_to_count' takes a numeric matrix 'exp.data' with only non-negative entries and normalizes each column (=> sample) to a fix number of counts.\cr
For every sample i, each feature j gets scaled to
 \deqn{exp.data[j, i] = (count * exp.data[j, i]) / sum(exp.data[, i])}
}
\examples{
some.matrix <- matrix(abs(rnorm(1000 * 5)), ncol = 5, nrow = 1000)
# each sample (=column) has different number of total counts:
apply(some.matrix, 2, sum)

normalized.matrix <- normalize_to_count(some.matrix)

# check:
apply(normalized.matrix, 2, sum)
}
