% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_gradient.R
\name{Trace.H.gradient}
\alias{Trace.H.gradient}
\title{Title}
\usage{
Trace.H.gradient(X = x_mat, Y = y_mat, C = c_mat, gamma.vec = v_vec)
}
\arguments{
\item{gamma.vec}{}
}
\value{
numeric list, same length as "gamma.vec"
}
\description{
Title
}
\examples{
Trace.H.gradient(tweak)
}
