% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_trainCorMod.R
\name{train_correlation_model}
\alias{train_correlation_model}
\title{Title}
\usage{
train_correlation_model(tweak, X.matrix, train.data.list, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
